Refactor: Reorganizes files so there are no repeated ORM config files
